+++ changelog.md
@@ -5,6 +5,7 @@
 - 🚧 Add new Album content type for sharing your pictures
 - Manage archives (in settings section)
 - Ability to configure python path (for media archiving)
+- Scheduled job to clean old or incomplete files from storage
 
 ## Changed